Cambie el int por Long

+++ src/main/java/com/joseAmbrocio/KinalApp/service/IProductoService.java
@@ -11,13 +11,13 @@ public interface IProductoService {
 
     Producto guardar(Producto producto);
 
-    Optional<Producto> buscarPorId(int codigoProducto);
+    Optional<Producto> buscarPorId(Long codigoProducto);
 
-    Producto actualizar(int codigoProducto, Producto producto);
+    Producto actualizar(Long codigoProducto, Producto producto);
 
-    void eliminar(int codigoProducto);
+    void eliminar(Long codigoProducto);
 
-    boolean existePorId(int codigoProducto);
+    boolean existePorId(Long codigoProducto);
 
     List<Producto> activo();
